requirements.txt
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
codegen==1.0
coreapi==2.3.3
coreschema==0.0.4
dateutils==0.6.6
decorator==4.4.0
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-filter==2.1.0
django-rest-auth==0.9.3
django-rest-framework-braces==0.3.4
django-rest-swagger==2.2.0
django-userforeignkey==0.3.0
djangorestframework==3.9.2
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Markdown==3.1
MarkupSafe==1.1.1
oauthlib==3.0.1
openapi-codec==1.3.2
parso==0.4.0
patterns==0.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
